drivers/mrf24j40: allow for basic self-test on init
During production it is often desirable for devices to perform
some kind of basic self-test to isolate defects.

For this it is necessary for the initialization not to hang if a
component is faulty / not connected.

This moves an already exising self-test that was previously enabled
as a debug option to an independent compile-time configurable.

It is necessary to call this in _init() before mrf24j40_hardware_reset()
as the reset function uses xtimer_usleep() which will cause another
thread to get scheduled.
If this thread (e.g. rpl or ipv6) then tries to access the netdev, RIOT
will crash.