gnrc_netif: add NETDEV_EVENT_PDU_CHANGED event
For some devices the maximum frame size is dependant on the configuration.

This adds a NETDEV_EVENT_PDU_CHANGED event that a network driver can generate
if a change in configuration resulted in a different PDU than before.

For this the driver must implement the NETOPT_MAX_PDU_SIZE `.get()` request.